requirements.txt
PyYAML
numpy
cv2
facenet_pytorch